Renaming package files to reflect attempted changed with the new wrapper class
